app/requirements.txt
python-dotenv
pandas
mecab
tweepy==4.1.0
markovify
pandas